Campaign container

+++ server/api/controllers/logs.controller.js
@@ -1,17 +1,29 @@
 import async from 'async';
 import * as _ from 'lodash';
 import mongoose from 'mongoose';
-import { buildLogUI } from '../../lib/logs';
+import { buildLogUI, buildNewLog } from '../../lib/logs';
 
 const logger = require('../../lib/logger.js')();
 
 const Log = mongoose.model('Log');
 
+function scribeSession(req) {
+  if (!req.cookies.scribe_session || !req.cookies.scribe_session.campaign) {
+    return false;
+  }
+  return {
+    campaignId: req.cookies.scribe_session.campaign,
+  };
+}
+
 export function getLogs(req, res) {
+  const authorized = scribeSession(req);
+  if (!authorized) { return res.status(403).send('Error: Forbidden'); }
+
   const { limit, skip } = req.query;
 
-  Log
-  .find({})
+  return Log
+  .find({ campaignId: authorized.campaignId })
   .skip(Number(skip) || 0)
   .limit(Number(limit) || 10)
   .sort({ time: -1 })
@@ -28,9 +40,11 @@ export function getLogs(req, res) {
 }
 
 export function getLog(req, res) {
+  if (!scribeSession(req)) { return res.status(403).send('Error: Forbidden'); }
+
   const { id } = req.params;
 
-  Log
+  return Log
   .findById(id)
   .lean()
   .exec((err, log) => {
@@ -46,9 +60,15 @@ export function getLog(req, res) {
 
 
 export function createLog(req, res) {
+  const authorized = scribeSession(req);
+
+  if (!authorized) { return res.status(403).send('Error: Forbidden'); }
+
+  const newLog = buildNewLog(req.body, authorized.campaignId);
+
   logger.log('createLog: %o', req.body);
 
-  Log.create(req.body, (err, log) => {
+  return Log.create(newLog, (err, log) => {
     if (err) {
       logger.log(`Error: ${err}`);
       return res.send(err);
@@ -60,9 +80,11 @@ export function createLog(req, res) {
 }
 
 export function updateLog(req, res) {
+  if (!scribeSession(req)) { return res.status(403).send('Error: Forbidden'); }
+
   const { id } = req.params;
 
-  Log.findByIdAndUpdate(id, req.body, { new: true }, (err, log) => {
+  return Log.findByIdAndUpdate(id, req.body, { new: true }, (err, log) => {
     if (err) {
       logger.log(`Error: ${err}`);
       return res.send(err);
@@ -73,9 +95,11 @@ export function updateLog(req, res) {
 }
 
 export function deleteLogs(req, res) {
+  if (!scribeSession(req)) { return res.status(403).send('Error: Forbidden'); }
+
   logger.log('deleteLogs: deleting logs:%j', req.body);
 
-  async.each(req.body, (logId, eachCallback) => {
+  return async.each(req.body, (logId, eachCallback) => {
     logger.log('deleting logs');
     Log.findByIdAndRemove(logId, {}, (err, log) => {
       logger.log(logId);
